TASK [custom/matrix-backup-s3 : ansible.builtin.include_tasks] ********************************* ERROR! We were unable to read either as JSON nor YAML, these are the errors we got from each: JSON: Expecting value: line 1 column 1 (char 0)
Syntax Error while loading YAML.
  did not find expected '-' indicator

The error appears to be in '/home/doof/gitrepo/matrix-docker-ansible-deploy/roles/custom/matrix-backup-s3/tasks/setup_install.yml': line 22, column 47, but may
be elsewhere in the file depending on the exact syntax problem.

The offending line appears to be:

  with_items:
    - {{matrix_backup_postgres_s3_identifier}}.service
                                              ^ here
We could be wrong, but this one looks like it might be an issue with
missing quotes. Always quote template expression brackets when they
start a value. For instance:

    with_items:
      - {{ foo }}

Should be written as:

    with_items:
      - "{{ foo }}"